:recycle: refactor(view): 重构 BottomSheet 状态管理
- 拆分稳定锚点与拖动、吸附运动状态，记录真实起点、目标、速度和 transitionId，支持动画中途改向并拒绝旧动画提交过期状态。
- 迁移业务调用方到 settledAnchor、motionState 和 awaitSettledAnchor，按语义方向协调课程避让、关闭监听及点击与返回拦截。
- 将 BottomSheet 组件整理到 ui.bottomsheet 包，拆分状态、peekHeight、Fling、NestedScroll 与 SceneStrategy 实现并补齐中文注释。
- 增加中途改向回归测试与 commonTest 依赖，同时更新 CMP 可复用组件索引。
- 验证 Desktop 测试、Android 与 iOS Simulator 编译通过；IDE 诊断及 git diff --check 无错误。
